# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for FindMe application.",
        "title": "FindMe API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/msg/add-user": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "An endpoint for adding users to a group chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Msg"
                ],
                "summary": "Add a user to a group chat",
                "parameters": [
                    {
                        "description": "Chat payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.AddUserChat"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "User Added",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "422": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    }
                }
            }
        },
        "/api/msg/delete-chat": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "An endpoint for deleting a group chat owned by the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Msg"
                ],
                "summary": "Delete a group chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "chat ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Chat deleted"
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    }
                }
            }
        },
        "/api/msg/delete-message": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "An endpoint for deleting a sent message of the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Msg"
                ],
                "summary": "Delete a sent message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Msg ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Message deleted"
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    }
                }
            }
        },
        "/api/msg/edit-message": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "An endpoint for editing a sent message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Msg"
                ],
                "summary": "Editing a sent message",
                "parameters": [
                    {
                        "description": "Message payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.EditMessage"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Message edited",
                        "schema": {
                            "$ref": "#/definitions/schema.DocMsgResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "403": {
                        "description": "permission denied",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "422": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    }
                }
            }
        },
        "/api/msg/leave-chat": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "An endpoint for leaving a group chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Msg"
                ],
                "summary": "Leave a group chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Chat removed"
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    }
                }
            }
        },
        "/api/msg/open-chat": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "An endpoint for opening a chat between users with IDs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Msg"
                ],
                "summary": "Open a chat between users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat opened",
                        "schema": {
                            "$ref": "#/definitions/schema.DocViewChatHistory"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "schema.AddUserChat": {
            "type": "object",
            "required": [
                "chat_id",
                "user_id"
            ],
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "schema.DocMsgResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "$ref": "#/definitions/schema.ViewMessage"
                }
            }
        },
        "schema.DocNormalResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "schema.DocViewChatHistory": {
            "type": "object",
            "properties": {
                "msg": {
                    "$ref": "#/definitions/schema.ViewChat"
                }
            }
        },
...
        "schema.EditMessage": {
            "type": "object",
            "required": [
                "msg",
                "msg_id"
            ],
            "properties": {
                "msg": {
                    "type": "string"
                },
                "msg_id": {
                    "type": "string"
                }
            }
        },
...
        "schema.ViewChat": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "string"
                },
                "group": {
                    "type": "boolean"
                },
                "message": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ViewMessage"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "schema.ViewMessage": {
            "type": "object",
            "properties": {
                "edited": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "sent": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
...
    }
}